requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
mangum==0.17.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
faker==20.1.0
pydantic==2.5.2
pytest==7.4.3
httpx==0.25.2
boto3==1.34.0
aws-lambda-powertools==2.30.0
aws-xray-sdk==2.12.1
docker==7.0.0
requests==2.31.0
pytest-docker==2.0.0
localstack==3.0.0
pytest-env==1.1.0
python-json-logger==2.0.7
moto==5.1.1 
